webapp/backend/requirements.txt
fastapi
uvicorn
transformers
torch
pydantic
python-jose
passlib[bcrypt]
python-multipart
